+++ app/app/src/main/java/ca/mabdu/fukidney/myapplication/StressApi.kt
@@ -15,7 +15,7 @@ interface StressApi {
 // These names MUST match the keys in your Cloudflare D1 database exactly
 data class StressReport(
     val ip_address: String,   // The unique ID (UUID)
-    val stress_level: Int,    // 1-5 from your ONNX model
+    val stress_level: Float,    // 1-5 from your ONNX model
     val department: String,
     val longitude: Double,    // From GPS
     val latitude: Double      // From GPS
